+++ src/pages/index.tsx
@@ -1,71 +1,105 @@
-import type { NextPage } from 'next';
-import Head from 'next/head';
-import Image from 'next/image';
-import styles from '../styles/Home.module.css';
+import React from 'react';
+import styled from '@emotion/styled';
 
-const Home: NextPage = () => {
+const Wrapper = styled.div``;
+const Title = styled.p`
+  @font-face {
+    font-family: 'Pretendard';
+    src: url('https://cdn.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff')
+      format('woff');
+    font-weight: 800;
+  }
+  width: 195px;
+  height: 70px;
+  margin: 0 auto;
+  margin-top: 42px;
+  text-align: center;
+  font-family: 'Pretendard';
+  //호진 TODO : 30px하면 다음줄로 넘어감
+  font-size: 29px;
+  color: #000000;
+`;
+
+const SubTitle = styled.p`
+  //호진 TODO : font 굵기가 약간 이상한거 같음
+  @font-face {
+    font-family: Apple SD Gothic Neo;
+    src: url('/fonts/AppleSDGothicNeoT.woff');
+    font-weight: 600;
+  }
+  width: 214px;
+  line-height: 16px;
+  font-size: 16px;
+  font-family: 'Apple SD Gothic Neo';
+  color: #514e4e;
+  margin: 0 auto;
+  margin-top: 22px;
+`;
+
+const CardContainer = styled.div`
+  width: 271px;
+  height: 335px;
+  box-shadow: 0px 10px 40px rgba(106, 106, 106, 0.25);
+  border-radius: 20px;
+  margin: 0 auto;
+  margin-top: 61.36px;
+  padding: 14px 0px;
+`;
+const CardFront = styled.div`
+  width: 239px;
+  height: 307px;
+  background-color: #f5f5f5;
+  border-radius: 13px;
+  margin: 0 auto;
+`;
+
+const CardImgContainer = styled.div`
+  // 호진 TODO : absolute로 적용하게 되면 다른 모바일 사이즈에서 안맞음 => 고민해봐야함
+  position: absolute;
+  top: 326px;
+  left: 566px;
+`;
+const CardQuestionImg = styled.img`
+  width: 61px;
+  height: 96px;
+  text-align: center;
+`;
+
+const StartButton = styled.button`
+  @font-face {
+    font-family: 'Pretendard';
+    src: url('https://cdn.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff')
+      format('woff');
+  }
+  display: block;
+  background-color: #fd6463;
+  width: 319px;
+  height: 60px;
+  border-radius: 36.5px;
+  border: none;
+  margin: 0 auto;
+  margin-top: 21px;
+  margin-bottom: 75px;
+  font-family: Pretendard;
+  font-size: 17px;
+  font-weight: 700;
+  letter-spacing: 0.02em;
+`;
+
+const Home = () => {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.tsx</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
+    <Wrapper>
+      <Title>나의 선물 유형을 찾아서</Title>
+      <SubTitle>내가 어떤 선물러인지 알아볼까요?</SubTitle>
+      <CardContainer>
+        <CardImgContainer>
+          <CardQuestionImg src="img/question.svg" alt="물음표" />
+        </CardImgContainer>
+        <CardFront />
+      </CardContainer>
+      {/* 호진 TODO : Button은 컴포넌트로 빼면 좋을듯 */}
+      <StartButton>테스트 시작하기</StartButton>
+    </Wrapper>
   );
 };
